Enemy LOS is now cleared using a Timer
Enemy now loses sight of player using a timer. Reversed engineered how I set this up with Blueprint in the past and set a timer using C++, was much easier than expected and works well enough. Need to make it clear the LastSeenLocation on the Blackboard after investigating and basic behaviour is looking great.